:bug: 3.6.7: Fix openai.py
